+++ src/main/java/blackjack/domain/Game.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.CardCount;
+import blackjack.domain.participant.DrawCount;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Player;
@@ -48,7 +48,7 @@ public class Game {
         player.stay();
     }
 
-    public CardCount drawDealerCards() {
+    public DrawCount drawDealerCards() {
         return dealer.drawCards(cardDeck);
     }
